+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Go service animation ([@TheMightyPenguin])
 * React service animation ([@TheMightyPenguin])
 * Our work page ([@duranmla])
+* Webpack base config for JS and SCSS ([@TheMightyPenguin])
 
 ### Fixed
 * Prevent content to broke after resolution change ([@duranmla])